+++ MyDrawingForm/IGraphics.cs
@@ -10,15 +10,15 @@ namespace MyDrawingForm
     public interface IGraphics
     {
         void ClearAll();
-        void DrawLine(float x1, float y1, float x2, float y2);
-        void DrawRectangle(float x, float y, float height, float width);
-        void DrawEllipse(float x, float y, float height, float width);
-        void DrawArc(float x, float y, float height, float width, float startAngle, float sweepAngle);
-        void DrawString(string text, float x, float y);
-        void DrawPolygon(float x, float y, float height, float width);
-        void DrawBoundingBox(float x, float y, float height, float width);
-        void DrawTextBoundingBox(float x, float y, float height, float width);
-        void DrawFilledEllipse(float x, float y, float height, float width);
+        void DrawLine(int x1, int y1, int x2, int y2);
+        void DrawRectangle(int x, int y, int height, int width);
+        void DrawEllipse(int x, int y, int height, int width);
+        void DrawArc(int x, int y, int height, int width, int startAngle, int sweepAngle);
+        void DrawString(string text, int x, int y);
+        void DrawPolygon(int x, int y, int height, int width);
+        void DrawBoundingBox(int x, int y, int height, int width);
+        void DrawTextBoundingBox(int x, int y, int height, int width);
+        void DrawFilledEllipse(int x, int y, int height, int width);
         void DrawDot(int x, int y, int height, int width);
     }
 }
